model_service/requirements.txt
fastapi==0.104.1
tensorflow==2.14.0
torch==2.1.0
transformers==4.34.0
xgboost==2.0.0
lightgbm==4.1.0
scikit-learn==1.3.1
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
statsmodels==0.14.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
